Rename android example app and package name
