+++ concordium-android-sdk-examples/app/src/main/java/com/concordium/example/wallet/Constants.kt
@@ -1,4 +1,4 @@
-package com.example.android_sdk_example
+package com.concordium.example.wallet
 
 import com.concordium.sdk.crypto.wallet.Network
