+++ concordium-android-sdk-examples/app/src/main/java/com/concordium/example/wallet/Storage.kt
@@ -1,4 +1,4 @@
-package com.example.android_sdk_example
+package com.concordium.example.wallet
 
 import android.content.Context
 import android.content.SharedPreferences
