+++ concordium-android-sdk-examples/app/src/main/java/com/concordium/example/wallet/activities/IdentityActivity.kt
@@ -1,4 +1,4 @@
-package com.example.android_sdk_example.activities
+package com.concordium.example.wallet.activities
 
 import android.content.Intent
 import android.os.Bundle
@@ -23,11 +23,11 @@ import com.concordium.sdk.transactions.CredentialRegistrationId
 import com.concordium.sdk.transactions.Index
 import com.concordium.sdk.transactions.TransactionExpiry
 import com.concordium.sdk.types.AccountAddress
-import com.example.android_sdk_example.Constants
-import com.example.android_sdk_example.Requests
-import com.example.android_sdk_example.Storage
-import com.example.android_sdk_example.services.ConcordiumClientService
-import com.example.android_sdk_example.ui.Container
+import com.concordium.example.wallet.Constants
+import com.concordium.example.wallet.Requests
+import com.concordium.example.wallet.Storage
+import com.concordium.example.wallet.services.ConcordiumClientService
+import com.concordium.example.wallet.ui.Container
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import org.apache.commons.codec.binary.Hex
 import java.util.Collections
