+++ concordium-android-sdk-examples/app/src/main/java/com/concordium/example/wallet/services/wallet_proxy/ProxyBackendConfig.kt
@@ -1,4 +1,4 @@
-package com.example.android_sdk_example.services.wallet_proxy
+package com.concordium.example.wallet.services.wallet_proxy
 
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import okhttp3.OkHttpClient
