+++ concordium-android-sdk-examples/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Android-sdk-example</string>
+    <string name="app_name">Android Wallet Example</string>
 </resources>
